Added an interface to deactivate the buffering on unaligned pages.
This is usefule, if PAGE_SIZE is to large for RAM.

If anyone would need to do it, without using so much RAM, you would
have to copy the data, to another flash page before.

Hower, this should then be implemented in another variation of
a GenericFlashDriver. Not inside the existing ones.